requirements.txt
python-dotenv
cerebras_cloud_sdk
gradio
apsw